Beowulf 3D Unity Project
A 3D action game simulating Beowulf's final battle against a dragon in
the epic poem of the same name. It was made for a class in Unity using
C#. The version we turned in added an attack animation and some UI
elements, but I didn't have access to that code, so this is the most
complete version prior. WASD or Arrow Keys control the character.